Theme JavaScript files - old 3msource@mmm intranet
These are the current 3msource javascript files loading through the theme.  GetAjaxContent is very old, insiteEditcheck is the code the puts a link at the top of the page to refresh the page and "inspect" the pages JSP portlets to get the XML data stored on the page.

jquery 3.1 is in there because the Theme_corpJS contains version 8 or 9.

Theme_corpJS has helper functions code for debugging, the theme js, and plugins for the theme / extensions of plugins.

ITCenter_SiteJS_AsIsInPreview is a minified version of a 2,000+ page site that was un-minified using http://jsbeautifier.org/.

It is all my custom JavaScript, JSON, Ajax, and jQuery code for that site.